backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
bcrypt==3.2.2
beautifulsoup4==4.14.3
boto3==1.42.75
botocore==1.42.75
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
ecdsa==0.19.2
exceptiongroup==1.3.1
fastapi==0.135.2
h11==0.16.0
idna==3.11
iniconfig==2.3.0
jmespath==1.1.0
mangum==0.21.0
packaging==26.0
passlib==1.7.4
pluggy==1.6.0
pyasn1==0.6.3
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.22
requests==2.33.0
rsa==4.9.1
s3transfer==0.16.0
six==1.17.0
soupsieve==2.8.3
starlette==1.0.0
tomli==2.4.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
